refactor(catalog): extract component read models
